requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
audioread=3.0.1=pypi_0
brotli=1.1.0=pypi_0
brotli-python=1.1.0=py311hdf8f085_1
bzip2=1.0.8=h10d778d_5
ca-certificates=2023.11.17=h8857fd0_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
decorator=5.1.1=pypi_0
et-xmlfile=1.1.0=pypi_0
fonttools=4.47.0=pypi_0
freetype=2.12.1=h60636b9_2
giflib=5.2.1=hb7f2c08_3
icu=73.2=hf5e326d_0
idna=3.6=pyhd8ed1ab_0
joblib=1.3.2=pypi_0
jpeg=9e=hb7f2c08_3
kiwisolver=1.4.5=pypi_0
lazy-loader=0.3=pypi_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libblas=3.9.0=20_osx64_openblas
libcblas=3.9.0=20_osx64_openblas
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.17=hb664fd8_1
libexpat=2.5.0=hf0c8a7f_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=13_2_0_h97931a8_1
libgfortran5=13.2.0=h2873a65_1
libhwloc=2.9.3=default_h24e0189_1009
libiconv=1.17=hd75f5a5_2
liblapack=3.9.0=20_osx64_openblas
libopenblas=0.3.25=openmp_hfef2a42_0
libpng=1.6.39=ha978bb4_0
libprotobuf=3.21.12=h7d26f99_2
librosa=0.10.1=pypi_0
libsqlite=3.44.2=h92b6c6a_0
libtiff=4.5.1=hcec6c5f_0
libwebp=1.3.2=hf6ce154_0
libwebp-base=1.3.2=h0dc2134_0
libxml2=2.11.6=hc0ae0f7_0
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=17.0.6=hb6ac08f_0
llvmlite=0.41.1=pypi_0
lz4=4.3.3=pypi_0
matplotlib=3.8.2=pypi_0
mkl=2022.2.1=h44ed08c_16952
msgpack=1.0.7=pypi_0
ncurses=6.4=h93d8f39_2
numba=0.58.1=pypi_0
numpy=1.26.3=pypi_0
openjpeg=2.5.0=h13ac156_2
openpyxl=3.1.2=pypi_0
openssl=3.2.0=hd75f5a5_1
packaging=23.2=pypi_0
pandas=1.5.3=pypi_0
pillow=10.0.1=pypi_0
pip=23.3.2=pyhd8ed1ab_0
platformdirs=4.1.0=pypi_0
pooch=1.8.0=pypi_0
pycparser=2.21=pypi_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.11.7=h9f0c242_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.11=4_cp311
pytorch=1.13.1=cpu_py311h66d45be_1
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0.1=pypi_0
readline=8.2=h9e318b2_1
requests=2.31.0=pyhd8ed1ab_0
rosbags=0.9.19=pypi_0
ruamel-yaml=0.18.5=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h6db0672_2
soundfile=0.12.1=pypi_0
soxr=0.3.7=pypi_0
tbb=2021.11.0=he51d815_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.13=h1abcd95_1
torch=1.13.1=pypi_0
torchvision=0.15.2a0=pypi_0
tqdm=4.66.1=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023d=h0c530f3_0
urllib3=2.1.0=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
xz=5.4.5=h6c40b1e_0
zlib=1.2.13=h8a1eda9_5
zstandard=0.22.0=pypi_0
zstd=1.5.5=h829000d_0
